+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Removed remaining `_type` field usages [#2017](https://github.com/ruflin/Elastica/pull/2017)
 ### Fixed
 * Fixed type-hint for `Elastica\Search::setOptionsAndQuery()` [#2009](https://github.com/ruflin/Elastica/pull/2009)
+* Fixed terms query params resolution [#2032](https://github.com/ruflin/Elastica/pull/2032)
 ### Security
 
 ## [7.1.2](https://github.com/ruflin/Elastica/compare/7.1.1...7.1.2)